separated commentlistslice from commentslice
